infrastructure/bridges/kafka-to-mqtt/requirements.txt
paho-mqtt==2.1.0
aiokafka==0.12.0